Real Initial Commit
This is the first real existance of GLRManager on Github. In this state the program is already set up with ImGUI + OpenGL GLFW as a GUI with some minor code to set up Configurations. cJSON has been properly hooked into the Library project but other requirements like cURL has yet to be properly incorperated.