Replaced cJSON library with simdjson. Refactrored Code to match the replacement.
Got rid of all cJSON mentions, generates a raw json file. Creation of jsons might be tougher, but loading jsons became so much more faster. Added Variables within the UserProfileManager to keep track of file paths and directories for ease of reference. Refactored code a little to use those new reference as well. Moved StartupAPPIDList as its really only used once and placed it within the loop instead.  It was only for startup so only will be called once