Updated Libraries to Latest Versions and Started work on a fix for SteamDB's Cloudflare
I also updated premake to get rid of the unessesary "Library" folder I was doing, unsure why I required it, it was unessesary and more or less a product of how i was taught Premake through college.

While I di dthat, I also updated Ultralight, Imgui and various helpers to latest versions, and started work on creating new functionality that uses the steam store itself to get and manage appIDs which is not protected by cloudflare.

My new methodology is to ask the user to locate a "main game" to search DLC from, as the AppID limit Greenluma forces one to make profiles based off games. More often CreamAPI would be far better at the functionality of unlocking DLC which is why there is now a generate creamAPI list so you can copy paste the DLC list to the creamapi.ini files.

So once I ask a user the main game, I'll call a function that uses that main game ID and Name to search and gather all the DLC assosiated with it.

I can make a few assumptions
1. You are using Greenluma to unlock DLC, therefore you probably own the base game anyways
2. Your using a manager to better switch between profiles and keep track of AppIDs
3. You probably don't wish to include free DLC in the list, nor are you aiming to consolidate multiple games within on elist

Number 3 can be worked around by simply selecting the next game and searching for DLC.

WIth those assumptions, I will now ask the user to search for game you wish to unlock DLC from, select it from a list, then search the DLCs for that game.

I may attempt to make that process automatic, as if you type the game in correctly, steam automatically makes the best result the first one in the list and more often than not is the game people would desire. I'll simply code in a function to overwrite the list with a new main game appid. Profiles would otherwise be unaffected.